ria_assistant_starter_kit/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
numpy==1.26.4
pandas==2.2.2
faiss-cpu==1.8.0.post1
rank-bm25==0.2.2
sentence-transformers==3.0.1
transformers==4.42.4
torch==2.3.1
scikit-learn==1.4.2
python-multipart==0.0.9
reportlab==4.2.0
pypdf==4.3.1
tqdm==4.66.4
streamlit==1.37.0
httpx==0.27.0
markdown-it-py==3.0.0
